feat: add StockCorrectionFixedItems
