+++ packages/apps/stock/src/screens/stockCorrections/StockCorrectionDetailsScreen.js
@@ -19,12 +19,10 @@
 import React, {useCallback, useEffect, useState} from 'react';
 import {ActivityIndicator} from 'react-native';
 import {
-  Button,
   HeaderContainer,
   PopUpOneButton,
   Screen,
   ScrollView,
-  useThemeColor,
 } from '@axelor/aos-mobile-ui';
 import {
   useDispatch,
@@ -32,18 +30,17 @@ import {
   useTranslator,
   HeaderOptionsMenu,
 } from '@axelor/aos-mobile-core';
-import {StockCorrectionHeader, StockCorrectionBody} from '../../components';
+import {
+  StockCorrectionHeader,
+  StockCorrectionBody,
+  StockCorrectionFixedItems,
+} from '../../components';
 import {fetchStockCorrectionReasons} from '../../features/stockCorrectionReasonSlice';
 import {fetchProductWithId} from '../../features/productSlice';
-import {
-  createCorrection,
-  updateCorrection,
-} from '../../features/stockCorrectionSlice';
 import {fetchProductIndicators} from '../../features/productIndicatorsSlice';
 import StockCorrection from '../../types/stock-corrrection';
 
 const StockCorrectionDetailsScreen = ({navigation, route}) => {
-  const Colors = useThemeColor();
   const I18n = useTranslator();
   const {loadingProduct, productFromId} = useSelector(state => state.product);
   const {activeCompany} = useSelector(state => state.user.user);
@@ -132,113 +129,6 @@ const StockCorrectionDetailsScreen = ({navigation, route}) => {
     initVariables();
   }, [initVariables]);
 
-  const handleSave = () => {
-    if (reason.id === null) {
-      // Required field
-      setPopUp(true);
-      return;
-    }
-
-    // Request AOS API
-    if (route.params.stockCorrection == null) {
-      // Stock correction doesn't exsist yet : creation
-      if (
-        stockProduct?.trackingNumberConfiguration == null ||
-        trackingNumber == null
-      ) {
-        dispatch(
-          createCorrection({
-            productId: stockProduct.id,
-            stockLocationId: stockLocation.id,
-            reasonId: reason.id,
-            status: StockCorrection.status.Draft,
-            realQty: realQty,
-          }),
-        );
-      } else {
-        dispatch(
-          createCorrection({
-            productId: stockProduct.id,
-            stockLocationId: stockLocation.id,
-            reasonId: reason.id,
-            trackingNumberId: trackingNumber.id,
-            status: StockCorrection.status.Draft,
-            realQty: realQty,
-          }),
-        );
-      }
-    } else {
-      // Stock correction already exists : update qty or reason
-      dispatch(
-        updateCorrection({
-          version: route.params.stockCorrection.version,
-          stockCorrectionId: route.params.stockCorrection.id,
-          realQty: realQty,
-          reasonId: reason.id,
-        }),
-      );
-    }
-    handleNavigation();
-  };
-
-  const handleValidate = () => {
-    if (reason.id === null) {
-      // Required field
-      setPopUp(true);
-      return;
-    }
-
-    // Request AOS API
-    if (route.params.stockCorrection == null) {
-      // Stock correction doesn't exsist yet : creation
-      if (
-        stockProduct?.trackingNumberConfiguration == null ||
-        trackingNumber == null
-      ) {
-        dispatch(
-          createCorrection({
-            productId: stockProduct.id,
-            stockLocationId: stockLocation.id,
-            reasonId: reason.id,
-            status: StockCorrection.status.Validated,
-            realQty: realQty,
-          }),
-        );
-      } else {
-        dispatch(
-          createCorrection({
-            productId: stockProduct.id,
-            stockLocationId: stockLocation.id,
-            reasonId: reason.id,
-            trackingNumberId: trackingNumber.id,
-            status: StockCorrection.status.Validated,
-            realQty: realQty,
-          }),
-        );
-      }
-    } else {
-      // Stock correction already exists : update qty or reason
-      dispatch(
-        updateCorrection({
-          version: route.params.stockCorrection.version,
-          stockCorrectionId: route.params.stockCorrection.id,
-          realQty: saveStatus ? null : realQty,
-          reasonId: saveStatus ? null : reason.id,
-          status: StockCorrection.status.Validated,
-        }),
-      );
-    }
-    handleNavigation();
-  };
-
-  const handleNavigation = () => {
-    if (route.params.externeNavigation === true) {
-      navigation.pop();
-    } else {
-      navigation.popToTop();
-    }
-  };
-
   const [popUp, setPopUp] = useState(false);
 
   React.useLayoutEffect(() => {
@@ -258,18 +148,18 @@ const StockCorrectionDetailsScreen = ({navigation, route}) => {
     <Screen
       removeSpaceOnTop={true}
       fixedItems={
-        <>
-          {saveStatus ? null : (
-            <Button
-              title={I18n.t('Base_Save')}
-              color={Colors.secondaryColor}
-              onPress={handleSave}
-            />
-          )}
-          {status === StockCorrection.status.Validated ? null : (
-            <Button title={I18n.t('Base_Validate')} onPress={handleValidate} />
-          )}
-        </>
+        <StockCorrectionFixedItems
+          navigation={navigation}
+          realQty={realQty}
+          reason={reason}
+          route={route}
+          saveStatus={saveStatus}
+          setPopUp={setPopUp}
+          status={status}
+          stockLocation={stockLocation}
+          stockProduct={stockProduct}
+          trackingNumber={trackingNumber}
+        />
       }
       loading={loadingProduct}>
       <HeaderContainer
